Change Reader#each to yield an ErrorRecord on error
Previously, the #each code would raise errors when a record could not be constructed (e.g., due
to a missing leader). This made it impossible for a user to catch an error and log/move on
within the block passed to Reader#each.

The code now yields an ErrorRecord that responds to #error, which simply returns the error
that cause the problem.

Use would be something like

     MARC::AlephSequential::Reader.new(filename).each do |r|
       next if MARC::AlephSequential::ErrorRecord === r
       puts r['245']
     end

Update README and spec to suit.